refactor(shikomi-gui): error.rs (643行) を error/mod.rs + error/tests.rs に分割 (Issue #96)
ペガサス指摘対応: 1ファイル500行制限超過を解消。

- error.rs (643行) を削除
- error/mod.rs (201行): 型定義・Serialize実装・helper関数
- error/tests.rs (441行): #[cfg(test)] mod tests の全テスト
- ipc_client/mod.rs の `pub mod error;` はそのまま（Rustが error/mod.rs を自動解決）
- cargo fmt 通過確認
